requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
faker>=18.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
flake8-docstrings>=1.7.0
flake8-bugbear>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pylint>=2.17.0

# Type stubs
types-python-dateutil>=2.8.0
types-cachetools>=5.3.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
sphinx-click>=4.4.0
myst-parser>=1.0.0

# Development tools
pre-commit>=3.0.0
ipython>=8.10.0
ipdb>=0.13.0

# Build tools
build>=0.10.0
twine>=4.0.0
wheel>=0.40.0
